Move ChainConfig to trinity.config, add node_class
