+++ tests/trinity/core/chain-management/test_initialize_database.py
@@ -5,12 +5,10 @@ from evm.db.backends.level import LevelDB
 from evm.db.chain import ChainDB
 
 from trinity.chains import (
+    ChainConfig,
     initialize_data_dir,
     initialize_database,
     is_database_initialized,
-)
-from trinity.utils.chains import (
-    ChainConfig,
 )
