+++ tests/trinity/core/chain-management/test_is_data_dir_initialized.py
@@ -4,10 +4,8 @@ import pytest
 from eth_utils import decode_hex
 
 from trinity.chains import (
-    is_data_dir_initialized,
-)
-from trinity.utils.chains import (
     ChainConfig,
+    is_data_dir_initialized,
 )
