+++ tests/trinity/core/chain-management/test_is_database_initialized.py
@@ -6,11 +6,9 @@ from evm.db.backends.level import LevelDB
 from evm.db.chain import ChainDB
 
 from trinity.chains import (
+    ChainConfig,
     initialize_data_dir,
     is_database_initialized,
-)
-from trinity.utils.chains import (
-    ChainConfig,
 )
